requirements/base.txt
# Base API
pytz==2018.7
python-slugify==1.2.6
Pillow==8.2.0
psycopg2==2.7.4 --no-binary psycopg2

# Django
django==3.1.0
django-crum==0.7.9

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.12.1
django-filter==2.4.0

# PyJWT
pyjwt==1.7.1

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise==5.2.0

# PyEasyToken
djangorestframework-simplejwt==4.4.0

# Celery
redis==3.2.0
django-redis==4.10.0
celery==4.2.1
flower==0.9.2
tornado>=5.0.0,<6.0.0